update stylelint config
