+++ packages/stylelint-config/package.json
@@ -49,8 +49,8 @@
     "execa": "^8.0.1",
     "fs-extra": "^11.2.0",
     "stylelint-config-clean-order": "^6.1.0",
-    "stylelint-config-html": "^1.1.0",
-    "stylelint-config-standard": "^36.0.1"
+    "stylelint-config-standard": "^36.0.1",
+    "stylelint-config-standard-scss": "^13.1.0"
   },
   "publishConfig": {
     "access": "public"
